third_party/WebKit/LayoutTests/xr/xrPresentationContext_context_creation.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
let width = 10;
let height = 10;

test(function() {
  let canvas = document.createElement("canvas");
  canvas.width = width;
  canvas.height = height;
  let ctx = canvas.getContext('xrpresent');
  assert_true(ctx instanceof XRPresentationContext);
}, "Test that canvas.getContext('xrpresent') returns an instance of XRPresentationContext");

</script>
